+++ wdio.conf.js
@@ -48,7 +48,7 @@ exports.config = {
   // NPM script (see https://docs.npmjs.com/cli/run-script) then the current working
   // directory is where your package.json resides, so `wdio` will be called from there.
   //
-  specs: ['./src/demo/examples/**/budget.feature'],
+  specs: ['./src/demo/examples/**/references.feature'],
   // Patterns to exclude.
   exclude: [
     // 'path/to/excluded/files'